+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/viewmodels/header/HeaderNavigationSpec.scala
@@ -16,8 +16,60 @@
 
 package uk.gov.hmrc.govukfrontend.views.viewmodels.header
 
+import play.api.libs.json.Json
 import uk.gov.hmrc.govukfrontend.views.viewmodels.JsonRoundtripSpec
-import uk.gov.hmrc.govukfrontend.views.viewmodels.header.Generators._
+import uk.gov.hmrc.govukfrontend.views.viewmodels.content.{Empty, HtmlContent, Text}
+import Generators._
 
-class HeaderNavigationSpec extends JsonRoundtripSpec[HeaderNavigation]
+class HeaderNavigationSpec extends JsonRoundtripSpec[HeaderNavigation] {
+  "HeaderNavigation" should {
+    "round trip html content successfully" in {
+      val testData =
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), HtmlContent("<strong>Html</strong>"))
 
+      val json = Json.toJson[HeaderNavigation](testData)
+
+      json.as[HeaderNavigation] should be(
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), HtmlContent("<strong>Html</strong>")))
+    }
+
+    "round trip empty content successfully" in {
+      val testData =
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), Empty)
+
+      val json = Json.toJson[HeaderNavigation](testData)
+
+      json.as[HeaderNavigation] should be(
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), Empty))
+    }
+
+    "round trip non-Empty content successfully" in {
+      val testData =
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), HtmlContent(""))
+
+      val json = Json.toJson[HeaderNavigation](testData)
+
+      json.as[HeaderNavigation] should be(
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), HtmlContent("")))
+    }
+
+    "move any text content to the content field" in {
+      val testData =
+        HeaderNavigation(Some("text"), Some("link"), active = true, Map("abc" -> "def"), Empty)
+
+      val json = Json.toJson[HeaderNavigation](testData)
+
+      json.as[HeaderNavigation] should be(
+        HeaderNavigation(None, Some("link"), active = true, Map("abc" -> "def"), Text("text")))
+    }
+
+    "ignore any text in the text field if the content field is non-empty" in {
+      val testData = HeaderNavigation(Some("text:text"), Some("link"), active = true, Map(), Text("content:text"))
+
+      val json = Json.toJson[HeaderNavigation](testData)
+
+      json.as[HeaderNavigation] should be(
+        HeaderNavigation(text = None, Some("link"), active = true, Map(), Text("content:text")))
+    }
+  }
+}